changed dir for cross agency script
